Added:  - unit tests for functions like push, pop. Also add few traces    for initAndClear unit test.
Refactor:
 - type of top in structure stack_t changed to size_t and
   properly refactor all methods.

+++ stack/inc/stack.h
@@ -2,11 +2,10 @@
 #define __STACK_H__
 
 #include <stddef.h>
-#include <stdint.h>
 
 typedef struct stack_t {
     void* __stackArray_p__;
-    int32_t __top__;
+    size_t __top__;
     size_t __capacity__;
     size_t __elementSize__;
 } stack_t;
